+++ test/test_reactivity.py
@@ -7,9 +7,7 @@ mol = Molecule("test/test_files/ethylene.xyz")
 
 
 def test_Distance():
-    """
-    Test Distance reactivity functionality
-    """
+    """Test Distance reactivity functionality."""
     d = reactivity.Distance(1, 2)
     s = d.get_settings(mol=mol)
 
@@ -17,9 +15,7 @@ def test_Distance():
 
 
 def test_Angle():
-    """
-    Test Angle reactivity functionality
-    """
+    """Test Angle reactivity functionality."""
     ang = reactivity.Angle(1, 2, 3)
     s = ang.get_settings(mol=mol)
 
@@ -27,9 +23,7 @@ def test_Angle():
 
 
 def test_Dihedral():
-    """
-    Test Angle reactivity functionality
-    """
+    """Test Angle reactivity functionality."""
     dihed = reactivity.Dihedral(1, 2, 3, 4)
     s = dihed.get_settings(mol=mol)