+++ test/test_utils.py
@@ -1,35 +1,9 @@
 from pymonad import curry
-from qmflows.utils import (chunksOf, concat, concatMap, zipWith, zipWith3)
-
-
-def test_chunksOf():
-    """
-    chunksOf([0, 1, 2, 3], 1) == [[0], [1], [2], [3]]
-    """
-    assert list(chunksOf([0, 1, 2, 3], 1)) == [[0], [1], [2], [3]]
-
-
-def test_concat():
-    """
-    Test list concatenation
-    """
-    xss = [[1], [2]]
-    assert concat(xss) == [1, 2]
-
-
-def test_concatMap():
-    """
-    concatMap f == concat (map f)
-    """
-    f = lambda x: [x * 2]
-    xs = [1, 2]
-    assert concatMap(f, xs) == list(concat(map(f, xs)))
+from qmflows.utils import zipWith, zipWith3
 
 
 def test_zipwith():
-    """
-    zipWith f xs ys ==  map (f) (zip  xs ys)
-    """
+    """zipWith f xs ys ==  map (f) (zip  xs ys)."""
     @curry
     def f(x, y):
         return x * y
@@ -41,9 +15,7 @@ def test_zipwith():
 
 
 def test_zipwith3():
-    """
-    zipWith f xs ys zs ==  map (f) (zip  xs ys zs)
-    """
+    """zipWith f xs ys zs ==  map (f) (zip  xs ys zs)."""
     @curry
     def f(x, y, z):
         return x * y + z
@@ -52,4 +24,5 @@ def test_zipwith3():
     ys = range(3, 6)
     zs = range(6, 9)
 
-    assert zipWith3(f)(xs)(ys)(zs) == [f(x, y, z) for (x, y, z) in zip(xs, ys, zs)]
+    assert zipWith3(f)(xs)(ys)(zs) == [f(x, y, z)
+                                       for (x, y, z) in zip(xs, ys, zs)]