+++ test/test_xyz_parser.py
@@ -1,18 +1,15 @@
+"""Test the xyz file format readers."""
 from qmflows.parsers import (manyXYZ, parse_string_xyz)
-from qmflows.utils import chunksOf
-import pytest
+from more_itertools import chunked
 
 
-@pytest.mark.slow
 def test_multiple_geometries():
-    """
-    Test the reading of multiples molecular geometries from a file.
-    """
+    """Test the reading of multiples molecular geometries from a file."""
     path_xyz = 'test/test_files/five_points_ethylene.xyz'
 
     with open(path_xyz, 'r') as f:
         ls = f.readlines()
 
-    xs = [''.join(x) for x in chunksOf(ls, 8)]
+    xs = [''.join(x) for x in chunked(ls, 8)]
 
     assert list(map(parse_string_xyz, xs)) == manyXYZ(path_xyz)